move fetchAccountAsync to common
