+++ packages/common/src/store/account/index.ts
@@ -1,4 +1,8 @@
 export { default as accountReducer, actions as accountActions } from './slice'
 export * as accountSelectors from './selectors'
-export { default as accountSagas } from './sagas'
+export {
+  default as accountSagas,
+  fetchAccountAsync,
+  cacheAccount
+} from './sagas'
 export * from './types'
