+++ packages/web/src/common/store/account/sagas.js
@@ -1,8 +1,7 @@
 import { userApiFetchSaga } from '@audius/common/api'
-import { ErrorLevel, Kind, Status } from '@audius/common/models'
+import { ErrorLevel, Kind } from '@audius/common/models'
 import {
   FeatureFlags,
-  recordIP,
   createUserBankIfNeeded,
   getRootSolanaAccount
 } from '@audius/common/services'
@@ -11,9 +10,10 @@ import {
   accountSelectors,
   cacheActions,
   profilePageActions,
-  chatActions,
   solanaSelectors,
-  getContext
+  getContext,
+  fetchAccountAsync,
+  cacheAccount
 } from '@audius/common/store'
 import { call, put, fork, select, takeEvery } from 'redux-saga/effects'
 
@@ -26,7 +26,6 @@ import { retrieveCollections } from '../cache/collections/utils'
 
 const { fetchProfile } = profilePageActions
 const { getFeePayer } = solanaSelectors
-const { fetchBlockees, fetchBlockers } = chatActions
 
 const {
   getUserId,
@@ -40,7 +39,6 @@ const {
 const {
   signedIn,
   showPushNotificationConfirmation,
-  fetchAccountSucceeded,
   fetchAccountFailed,
   fetchAccount,
   fetchLocalAccount,
@@ -49,11 +47,8 @@ const {
   tikTokLogin,
   fetchSavedPlaylists,
   addAccountPlaylist,
-  resetAccount,
-  setWalletAddresses
+  resetAccount
 } = accountActions
-
-const IP_STORAGE_KEY = 'user-ip-timestamp'
 
 /**
  * Sets the sentry user so that alerts are tied to a user
@@ -74,26 +69,6 @@ const setSentryUser = (sentry, user, traits) => {
       ...traits
     })
   })
-}
-
-function* recordIPIfNotRecent(handle) {
-  const audiusBackendInstance = yield getContext('audiusBackendInstance')
-  const localStorage = yield getContext('localStorage')
-  const timeBetweenRefresh = 24 * 60 * 60 * 1000
-  const now = Date.now()
-  const minAge = now - timeBetweenRefresh
-  const storedIPStr = yield call([localStorage, 'getItem'], IP_STORAGE_KEY)
-  const storedIP = storedIPStr && JSON.parse(storedIPStr)
-  if (!storedIP || !storedIP[handle] || storedIP[handle].timestamp < minAge) {
-    const { userIP, error } = yield call(recordIP, audiusBackendInstance)
-    if (!error) {
-      yield call(
-        [localStorage, 'setItem'],
-        IP_STORAGE_KEY,
-        JSON.stringify({ ...storedIP, [handle]: { userIP, timestamp: now } })
-      )
-    }
-  }
 }
 
 // Tasks to be run on account successfully fetched, e.g.
@@ -181,80 +156,6 @@ function* onSignedIn({ payload: { account } }) {
   }
 }
 
-export function* fetchAccountAsync({ isSignUp = false }) {
-  const remoteConfigInstance = yield getContext('remoteConfigInstance')
-  const authService = yield getContext('authService')
-  const audiusBackendInstance = yield getContext('audiusBackendInstance')
-
-  const accountStatus = yield select(accountSelectors.getAccountStatus)
-
-  // Don't revert successful local account fetch
-  if (accountStatus !== Status.SUCCESS) {
-    yield put(accountActions.fetchAccountRequested())
-  }
-
-  const { accountWalletAddress: wallet, web3WalletAddress } = yield call([
-    authService,
-    authService.getWalletAddresses
-  ])
-
-  if (!wallet) {
-    yield put(
-      fetchAccountFailed({
-        reason: 'ACCOUNT_NOT_FOUND'
-      })
-    )
-    return
-  }
-
-  const accountData = yield call(userApiFetchSaga.getUserAccount, {
-    wallet
-  })
-
-  if (!accountData || !accountData.user) {
-    yield put(
-      fetchAccountFailed({
-        reason: 'ACCOUNT_NOT_FOUND'
-      })
-    )
-    return
-  }
-  const account = accountData.user
-
-  if (account.is_deactivated) {
-    yield put(accountActions.resetAccount())
-    yield put(
-      fetchAccountFailed({
-        reason: 'ACCOUNT_DEACTIVATED'
-      })
-    )
-    return
-  }
-
-  // Set the userId in the remoteConfigInstance
-  remoteConfigInstance.setUserId(account.user_id)
-
-  yield call(recordIPIfNotRecent, account.handle)
-
-  // Cache the account and put the signedIn action. We're done.
-  yield call(cacheAccount, account)
-  yield put(
-    setWalletAddresses({ currentUser: wallet, web3User: web3WalletAddress })
-  )
-
-  // Sync current user info to libs
-  const libs = yield call([
-    audiusBackendInstance,
-    audiusBackendInstance.getAudiusLibs
-  ])
-  yield call([libs, libs.setCurrentUser], {
-    wallet,
-    userId: account.user_id
-  })
-
-  yield put(signedIn({ account, isSignUp }))
-}
-
 export function* fetchLocalAccountAsync() {
   const localStorage = yield getContext('localStorage')
 
@@ -273,33 +174,6 @@ export function* fetchLocalAccountAsync() {
     yield put(fetchAccountFailed({ reason: 'ACCOUNT_NOT_FOUND_LOCAL' }))
   }
 }
-
-function* cacheAccount(account) {
-  const localStorage = yield getContext('localStorage')
-  const collections = account.playlists || []
-
-  yield put(
-    cacheActions.add(Kind.USERS, [
-      { id: account.user_id, uid: 'USER_ACCOUNT', metadata: account }
-    ])
-  )
-
-  const formattedAccount = {
-    userId: account.user_id,
-    collections
-  }
-
-  yield call([localStorage, 'setAudiusAccount'], formattedAccount)
-  yield call([localStorage, 'setAudiusAccountUser'], account)
-
-  yield put(fetchAccountSucceeded(formattedAccount))
-
-  // Fetch user's chat blockee and blocker list after fetching their account
-  yield put(fetchBlockees())
-  yield put(fetchBlockers())
-}
-
-// Pull from redux cache and persist to local storage cache
 export function* reCacheAccount() {
   const localStorage = yield getContext('localStorage')
   const account = yield select(getAccountToCache)
@@ -457,7 +331,7 @@ function* watchAddAccountPlaylist() {
 }
 
 function* watchResetAccount() {
-  yield takeEvery(resetAccount.type, function* (action) {
+  yield takeEvery(resetAccount.type, function* () {
     const audiusBackendInstance = yield getContext('audiusBackendInstance')
     const localStorage = yield getContext('localStorage')
     const libs = yield call(audiusBackendInstance.getAudiusLibs)
