+++ packages/web/src/common/store/pages/signon/sagas.ts
@@ -34,7 +34,8 @@ import {
   getContext,
   confirmerActions,
   confirmTransaction,
-  getSDK
+  getSDK,
+  fetchAccountAsync
 } from '@audius/common/store'
 import {
   Genre,
@@ -59,7 +60,7 @@ import {
   takeLatest
 } from 'typed-redux-saga'
 
-import { fetchAccountAsync, reCacheAccount } from 'common/store/account/sagas'
+import { reCacheAccount } from 'common/store/account/sagas'
 import { identify, make } from 'common/store/analytics/actions'
 import * as backendActions from 'common/store/backend/actions'
 import { retrieveCollections } from 'common/store/cache/collections/utils'
@@ -666,7 +667,7 @@ function* signUp() {
         // TODO (PAY-3479): This is temporary until hedgehog is fully moved out of libs
         yield* call(refreshHedgehogWallet)
         yield* put(signOnActions.sendWelcomeEmail(name))
-        yield* fetchAccountAsync({ isSignUp: true })
+        yield* call(fetchAccountAsync, { isSignUp: true })
         yield* put(signOnActions.followArtists())
         yield* put(make(Name.CREATE_ACCOUNT_COMPLETE_CREATING, { handle }))
         yield* put(signOnActions.signUpSucceeded())
@@ -738,7 +739,7 @@ function* repairSignUp() {
           console.info('Successfully repaired user')
           yield* put(make(Name.SIGN_UP_REPAIR_SUCCESS, {}))
           yield* put(signOnActions.sendWelcomeEmail(metadata.name))
-          yield* fetchAccountAsync({ isSignUp: true })
+          yield* call(fetchAccountAsync, { isSignUp: true })
         },
         function* ({ timeout }) {
           const reportToSentry = yield* getContext('reportToSentry')
